+++ tests/test_utils.py
@@ -856,6 +856,25 @@ class BuildLookupFromCSVTests(unittest.TestCase):
         self.assertEqual(lookup_dict[4]['header3'], 'bar')
         self.assertEqual(lookup_dict[1]['header1'], 1)
 
+    def test_csv_utf8_encoding(self):
+        """utils: test that CSV read correctly with UTF-8 encoding."""
+        from natcap.invest import utils
+
+        csv_file = os.path.join(self.workspace_dir, 'csv.csv')
+        with open(csv_file, 'w', encoding='utf-8') as file_obj:
+            file_obj.write(textwrap.dedent(
+                """
+                header1,HEADER2,header3
+                1,2,bar
+                4,5,FOO
+                """
+            ).strip())
+        lookup_dict = utils.build_lookup_from_csv(
+            csv_file, 'header1')
+        self.assertEqual(lookup_dict[4]['header2'], 5)
+        self.assertEqual(lookup_dict[4]['header3'], 'foo')
+        self.assertEqual(lookup_dict[1]['header1'], 1)
+
     def test_csv_utf8_bom_encoding(self):
         """utils: test that CSV read correctly with UTF-8 BOM encoding."""
         from natcap.invest import utils
@@ -901,6 +920,22 @@ class BuildLookupFromCSVTests(unittest.TestCase):
         self.assertEqual(lookup_dict[4]['header 2'], 5)
         self.assertEqual(lookup_dict[4]['header 3'], 'foo')
         self.assertEqual(lookup_dict[1]['header 1'], 1)
+
+    def test_csv_error_non_utf8_character(self):
+        """utils: test that error is raised on non-UTF8 character."""
+        from natcap.invest import utils
+
+        csv_file = os.path.join(self.workspace_dir, 'csv.csv')
+        with codecs.open(csv_file, 'w', encoding='iso-8859-1') as file_obj:
+            file_obj.write(textwrap.dedent(
+                """
+                header 1,HEADER 2,header 3
+                1,2,bar1
+                4,5,FÖÖ
+                """
+            ).strip())
+        with self.assertRaises(UnicodeDecodeError):
+            utils.build_lookup_from_csv(csv_file, 'header 1')
 
     def test_expand_path(self):
         """utils: test path expansion function."""