+++ workbench/src/main/baseUrl.js
@@ -1,5 +1,5 @@
 import ELECTRON_DEV_MODE from './isDevMode';
 
 export default (ELECTRON_DEV_MODE)
-  ? 'http://127.0.0.1:3000/'
+  ? 'http://localhost:5173/' // default port for vite dev server
   : `file://${__dirname}/../`; // resolves to the build/ dir