+++ workbench/src/renderer/menubar/handlers.js
@@ -1,7 +1,6 @@
 import { ipcMainChannels } from '../../main/ipcMainChannels';
 
 const { ipcRenderer } = window.Workbench.electron;
-const { LOGFILE_PATH } = window.Workbench;
 
 export function handleClickExternalURL(event) {
   event.preventDefault();
@@ -13,6 +12,6 @@ export function handleClickExternalURL(event) {
 export function handleClickFindLogfiles() {
   ipcRenderer.send(
     ipcMainChannels.SHOW_ITEM_IN_FOLDER,
-    LOGFILE_PATH
+    window.Workbench.ELECTRON_LOG_PATH,
   );
 }